Remove unsafe, eliminate panics, and split explorer page
Address OSS-readiness quality issues from #46.

Emergency fixes:
- Replace the raw-pointer `unsafe` callback in FileListDelegate with an
  Arc<dyn Fn> cloned per row, removing the only `unsafe` block.
- Move open_preview file I/O off the UI thread via background_spawn, applying
  the result through spawn_in/update_in.
- Make SearchService initialization non-fatal: the app starts with search
  disabled instead of panicking, and engine setup propagates the home-dir
  lookup error instead of calling expect().
- Replace the hand-rolled date math in format_date with time's
  OffsetDateTime, fixing wrong month/day for most dates.
- Harden the mdfind invocation (-onlyin/-literal, exit-code handling).
- Surface directory-listing and search errors in the footer status bar
  instead of only logging them.
- Drop the duplicate apply_filter() call in close_search.

File decomposition:
- Split the 688-line explorer/mod.rs into state, navigation, preview, search,
  and list_setup modules; mod.rs now only wires modules and trait impls.
- Collect magic numbers (window/column sizes, thresholds, default limits) into
  a new core::config module.

Release Notes:

- Fixed incorrect modification dates shown in the file list
- Improved error reporting by surfacing file and search errors in the status bar
