+++ src/core/mod.rs
@@ -1,2 +1,3 @@
+pub mod config;
 pub mod errors;
 pub mod telemetry;
